AMD64 - AUDIT RUN - Fix format strings, size_t, and other issues

+++ contrib/smbfs/lib/smb/nls.c
@@ -41,6 +41,7 @@
 #endif
 #include <errno.h>
 #include <stdio.h>
+#include <string.h>
 #include <strings.h>
 #include <stdlib.h>
 #include <locale.h>
@@ -125,7 +126,7 @@ char *
 nls_str_toloc(char *dst, const char *src)
 {
 	char *p = dst;
-	int inlen, outlen;
+	size_t inlen, outlen;
 
 	if (!iconv_loaded)
 		return strcpy(dst, src);
@@ -147,7 +148,7 @@ char *
 nls_str_toext(char *dst, const char *src)
 {
 	char *p = dst;
-	int inlen, outlen;
+	size_t inlen, outlen;
 
 	if (!iconv_loaded)
 		return strcpy(dst, src);
@@ -170,7 +171,7 @@ nls_mem_toloc(void *dst, const void *src, int size)
 {
 	char *p = dst;
 	const char *s = src;
-	int inlen, outlen;
+	size_t inlen, outlen;
 
 	if (!iconv_loaded)
 		return memcpy(dst, src, size);
@@ -195,7 +196,7 @@ nls_mem_toext(void *dst, const void *src, int size)
 {
 	char *p = dst;
 	const char *s = src;
-	int inlen, outlen;
+	size_t inlen, outlen;
 
 	if (size == 0)
 		return NULL;
